backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-ai-slim[google]
python-dotenv
praw
pytrends
